Add Tailwindcss 2.0
Used compatibility mode for PostCSS 7, as otherwise was not
working for PostCSS 8. Will need to investigate this.

Followed this guide (https://rubyyagi.com/tailwindcss2-rails6/) and the
docs (https://tailwindcss.com/docs/installation). Still not 100% on
what every step does, so will need to come back and understand it,
especially if we run into issues :).